+++ yash-env/src/subshell.rs
@@ -385,7 +385,7 @@ mod tests {
     #[test]
     fn jobs_disowned_in_subshell() {
         in_virtual_system(|mut env, _state| async move {
-            let index = env.jobs.add(Job::new(Pid(123)));
+            let index = env.jobs.insert(Job::new(Pid(123)));
             let subshell = Subshell::new(move |env, _job_control| {
                 Box::pin(async move { assert!(!env.jobs[index].is_owned) })
             });